chore: improve semantics and add rustfmt rules
